Added support to get rid of Phix -or similar- used during the sequencing.
User should give as input parameter the Phix sequence.
Phix sequence will split into MINLENGT fragments to see if the first part of the read maps to any fragment.
If that happens, the complete trimmed read is align against the whole Phix sequence.
Reads which map to the this sequence are removed from downstream analyses.